Dash_GCP_largesize/requirements.txt
dash==1.11.0
dash-core-components==1.9.1
dash-html-components==1.0.3
dash-renderer==1.4.0
dash-table==4.6.2
tensorflow==2.1.0
numpy==1.18.3
Pillow==7.1.2